requirements.txt
contourpy==1.0.5
cycler==0.11.0
fonttools==4.38.0
importlib-metadata==5.0.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.1
numba==0.56.3
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==5.4.1
scipy==1.9.3
six==1.16.0
zipp==3.10.0
